Locate methods in right folders, GetScorer method added to FuzzyMatch file as it's private
