+++ Search_Engine/Compute/Extract/ExtractAll.cs
