+++ Search_Engine/Compute/Extract/ExtractOne.cs
