+++ Search_Engine/Compute/Extract/ExtractSorted.cs
